deflake more tests
